Update MongoDB week 3
During the last iteration in London we opted to deviate from the syllabus as we
decided that it was too advanced for the students. This was what we came up
with instead, and feel it's a better lesson overall given the amount of
experience of MongoDB that students will have at this point in the course.

We appreciate this does not fall into the same format as the rest of the
syllabus, however we feel it is a good starting point to work from.

Co-authored-by: Greg Dyke <gregdyke@gmail.com>