Remove a couple of more advanced coding standards
We don't teach our learners about testing, so we shouldn't tell them to
have test coverage.

We explicitly teach them about setting style properties, so it's
probably confusing to tell them not to use them while we're teaching
them. It feels like we really want a document which grows
module-by-module, with each stage covering just what we've taught
them...

But if we're going to link to this in every week's repo, we should
probably remove at least these things.