JS-3 Week 2: Replace the TfL API exercise with an easier one
It's been a recurring comment from trainees and TAs that the TfL API
exercise is confusing because the API is complex and the documentation
is a bit obscure for someone who sees a REST API for the first time.

Replace it with an easier exercise that uses the Chuck Norris Jokes API
(https://api.chucknorris.io/).