DB: Fix code examples: A request handler must always send a response
The code examples of Week 2 and Week 3 of the DB module handle database
errors incorrectly whereby they only log the error to the console and
do not send a response. This is bad practice as it would leave the
client connection hanging until client timeout. A request handler must
always send a response.