Modify HTML/CSS Wk4 lesson to not use removed deployment workshop
The deployment workshop has been removed in favour of a guide, so change to inline instructions to deploy coursework projects. It's also worth noting that the workshop instructed trainees to deploy the cakes site, which isn't started until the week 4 coursework (i.e. before this class). I've switched it to the Bikes for Refugees project (week 1 coursework) instead