Update lessons to remove most dependency on Glitch
The current lessons have a heavy dependency on Glitch, with most exercises asking the trainees to do their work on Glitch.
This has caused some confusion among the trainees as they were wondering the importance of Glitch and how it relates to Node.

The following changes move Glitch more into the background and ask trainees to do most of their work locally, with the usual Pull Request procedure to submit their homework solutions.

Main changes in short:
* Moved all of the previous Glitch projects to https://github.com/CodeYourFuture/Node-Exercises
* Changed all of the notes that ask for remixes to simply clone the above repo and do the changes over there
* Updated Week 1's intro to ask trainees to play with the project locally instead of remixing it on Glitch
* Changed the demo project from `cyfcats` to `cyf-simple-server`, as the latter is more relevant in relation to API creation.